frontend/pyproject.toml
[project]
name = "frontend"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "cloudflare>=3.1.1",
    "disnake>=2.10.1",
    "python-dotenv>=1.0.1",
]
